refactor: musical-service 폴더구조 도메인 방식 변경
