+++ musical/src/main/java/com/truve/platform/musical/show/domain/entity/Artist.java
@@ -1,4 +1,4 @@
-package com.truve.platform.show.service.domain.entity;
+package com.truve.platform.musical.show.domain.entity;
 
 import com.truve.platform.common.support.BaseEntity;
