+++ musical/src/main/java/com/truve/platform/musical/show/domain/entity/ShowSchedule.java
@@ -1,9 +1,9 @@
-package com.truve.platform.show.service.domain.entity;
+package com.truve.platform.musical.show.domain.entity;
 
 import java.time.LocalDateTime;
 
 import com.truve.platform.common.support.BaseEntity;
-import com.truve.platform.show.service.domain.constant.ShowScheduleStatus;
+import com.truve.platform.musical.show.domain.constant.ShowScheduleStatus;
 
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
