+++ musical/src/main/java/com/truve/platform/musical/show/repository/ShowScheduleRepository.java
@@ -1,4 +1,4 @@
-package com.truve.platform.show.service.repository;
+package com.truve.platform.musical.show.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.truve.platform.show.service.domain.entity.ShowSchedule;
+import com.truve.platform.musical.show.domain.entity.ShowSchedule;
 
 public interface ShowScheduleRepository extends JpaRepository<ShowSchedule, Long> {
